transfer sound logic from child to parent
